requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
Pygments==2.18.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
